Major improvements to the filtering code
The code is now pretty slick, I think.  Adding a new filter that is
similar to one that's already implemented involves just added a few
lines of code, as I tested with the TimesReviewedFilter.  It took me a
little longer tonight than I was hoping, but I'm really happy with the
results.