Some work on getting fixtures working right.
I should be ready to put in irregular forms and have it work.

The point of the fixture is just for testing, really.  I need to figure
out how to make syncdb not load the fixture by default, but the tests
to.  The real "fixture" should be in add_to_db.py, because that can be
run regardless of the pk's, whereas the fixture has pks in it, and they
could conflict if you have a fixture for each language.