Trying to improve add_to_db
In particular, I would like to make the process of adding irregular
forms for testing easier than it has been, and also make it persistent
so that when I want to make a default database, it is nice and easy.  So
I've restructured add_to_db.py a bit to make it handle adding words and
irregular forms and all of that.  As I've been doing this, I think a
smarter thing would be to have a json file or set of json files
describing all of these objects, then load that json file to add things
to the database.  That would be easier than writing it in python.  But
at least this is a start.