add_to_db.py is now much cooler
I made it pretty much language agnostic, so that it could feasibly be
used for other languages with only very slight modification.  The main
deal now is language_data.json, where all of the information needed for
the language is stored in a self-contained location in a relatively
compact format.  add_to_db.py reads the json file and adds whatever it
sees.  And there's an option to create a fixture file that's used for
testing purposes, so fixtures/initial_data.json will change shortly.