Added aorist and perfect forms for ἵστημι
Sadly, a lot of the perfect forms had to be entered as irregular.  Oh
well.  If I come across another verb with similar behavior, it may be
worth it to add code for it, but until then this will work.  But this
finishes the initial -μι verbs introduced in Hansen and Quinn.  All
forms are there.