Finished with the test code cleanup
What I wanted to do, with putting the Conjugation instantiation into a
separate method, turned out to be about 3 times slower, so I gave up on
it.  So this was mostly just adding in a default tearDown to reset
conjugation.verbose to False, and localizing the imperative test cases
like that regular cases.