Mostly this makes logging in and out better
There are some other changes, like making the word-adding functionality
actually work.  But because things are in such a state of flux, I'm not
going to be as careful with my commits as I normally am.  I'm just
committing everything that has changed, and I'll fix what needs to be
fixed later.