Switching from "difficulty" to "memory index"
That is, try to keep track of how long a particular word has been in your
memory, and slowly increase the time between reviews as you keep on getting it
right.  I think that's a better model for helping you review.  It's not totally
integrated as I want it yet, but the basics seem to be working.